+++ src/components/loan/CustomizeCard/CustomizeCardTypes.ts
@@ -1,0 +1,15 @@
+export type CustomizeCardRef = {
+  scrollToForm: () => void
+}
+
+export type FormValues = {
+  amount: number
+  lastName: string
+  firstName: string
+  patronymic?: string
+  term: string
+  email: string
+  birthdate: string
+  passportSeries: string
+  passportNumber: string
+}
